requirements.txt
# API & LLM
anthropic>=0.32.0
google-generativeai>=0.5.0

# Chess engine
chess>=1.10.0
numpy>=1.24.0

# Memory & Vector DB
sentence-transformers>=2.2.0

# Database & ORM
sqlalchemy>=2.0.0

# Web server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-socketio>=5.11.0
python-engineio>=4.9.0
jinja2>=3.1.0
aiofiles>=23.0.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
